Move monitoring stuff into own subpackage
